+++ L2J_Mobius_CT_2.6_HighFive/java/com/l2jmobius/gameserver/model/actor/instance/L2PcInstance.java
@@ -6715,7 +6715,7 @@ public final class L2PcInstance extends L2Playable
 					player.getStat().setExp(rset.getLong("exp"));
 					player.setExpBeforeDeath(rset.getLong("expBeforeDeath"));
 					player.getStat().setLevel(rset.getByte("level"));
-					player.getStat().setSp(rset.getInt("sp"));
+					player.getStat().setSp(rset.getLong("sp"));
 					
 					player.setWantsPeace(rset.getInt("wantspeace"));
 					
@@ -7031,7 +7031,7 @@ public final class L2PcInstance extends L2Playable
 					subClass.setClassId(rs.getInt("class_id"));
 					subClass.setLevel(rs.getByte("level"));
 					subClass.setExp(rs.getLong("exp"));
-					subClass.setSp(rs.getInt("sp"));
+					subClass.setSp(rs.getLong("sp"));
 					subClass.setClassIndex(rs.getInt("class_index"));
 					
 					// Enforce the correct indexing of _subClasses against their class indexes.
